organized a bit for tomorrow
