+++ db_operations.py
@@ -11,9 +11,6 @@ def insert_data(data_list):
 
     with connect() as mydb:
         connection = mydb.cursor()
-        # SQL insert statement
         sql = "INSERT INTO data (title, company, location, management_level, salary, tryb_pracy, etat, kontrakt, specjalizacja) VALUES (%s,%s,%s,%s,%s,%s,%s,%s,%s)"
-
-        # Execute the SQL statement
         connection.executemany(sql, data_list)
         mydb.commit()
